requirements.txt
Django==5.1.5
djangorestframework==3.15.2
streamlit==1.41.1
python-decouple==3.8
celery==5.4.0
redis==5.2.1
langchain==0.3.16
langchain-openai==0.3.2
langchain-anthropic==0.3.5
langchain-experimental==0.3.4
langgraph==0.2.68
openpyxl==3.1.5
qdrant-client==1.13.2
qdrant-client[fastembed]>=1.8.2
python-dotenv==1.0.1
langfuse==2.58.0
mysqlclient==2.2.7
tabulate==0.9.0
